Status Bar & Properties
Created Status bar blueprint
Added a status bar to the BaseAgent Blueprint that visualises the agents
current health and energy
Added methods for killing and damaging agents
Added death state to agent and made it so that it ragdolls on death.
Added ondeath event that can be used to destroy any unused components.
Added properties for detailing an agents viewing arc. Currently has
field of view and viewing distance so far.
set energy deteriation to 0.1 per second.